+++ src/main/java/com/tagmycode/plugin/gui/AbstractSnippetsListGui.java
@@ -3,5 +3,5 @@ package com.tagmycode.plugin.gui;
 import javax.swing.*;
 
 public abstract class AbstractSnippetsListGui extends AbstractGui implements ISnippetsListGui, IAbstractGUI {
-    public abstract JComponent getSnippetsComponent();
+    public abstract JComponent getJTable();
 }