+++ src/main/java/com/tagmycode/plugin/operation/DeleteSnippetOperation.java
@@ -1,12 +1,12 @@
 package com.tagmycode.plugin.operation;
 
-import com.tagmycode.plugin.gui.form.SnippetsTab;
+import com.tagmycode.plugin.gui.form.SnippetsPanel;
 import com.tagmycode.sdk.model.Snippet;
 
 public class DeleteSnippetOperation extends AbstractSnippetOperation {
     private Snippet snippet;
 
-    public DeleteSnippetOperation(SnippetsTab snippetsTab, Snippet snippet) {
+    public DeleteSnippetOperation(SnippetsPanel snippetsTab, Snippet snippet) {
         super(snippetsTab.getFramework(), snippetsTab);
         this.snippet = snippet;
     }