+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - added title and description to snippet view
 - added log4j.properties configuration
 - log error if can't open web browser
+- text can be dragged into table to create a new snippet
+- snippets can be dragged directly into editor and the code are copied
 
 ### Fixed
 - show a dialog error if verification code is not correct